+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.0
-* added the `BackupImport` class.
+* added the `BackupImport` class;
+* by default the plugin uses the `APP/backups` directory to save the backups files.
 
 ## 1.0 branch
 ### 1.0.3